refactor: rm circular module dependencies
