+++ src/components/surface/index.ts
@@ -1,4 +1,5 @@
-export { default as Surface, useSurface } from './surface';
+export { default as Surface } from './surface';
+export { useSurface } from './surface.context';
 export { surfaceClassNames } from './surface.styles';
 export type {
   SurfaceBackgroundProps,
