+++ src/components/toast/toast.tsx
@@ -14,8 +14,8 @@ import { AnimationSettingsProvider } from '../../helpers/internal/contexts';
 import type { ViewRef } from '../../helpers/internal/types';
 import { createContext } from '../../helpers/internal/utils';
 import * as ToastPrimitive from '../../primitives/toast';
-import type { ToastComponentProps } from '../../providers/toast';
 import { useToastConfig } from '../../providers/toast/toast-config.context';
+import type { ToastComponentProps } from '../../providers/toast/types';
 import { Button } from '../button';
 import { flattenColorOverBackground } from '../glass-view/glass-view.utils';
 import { useToastRootAnimation } from './toast.animation';
